requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
blessed==1.20.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
cssselect2==0.7.0
daphne==4.1.2
defusedxml==0.7.1
Django==3.2
django-cors-headers==4.4.0
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.12.4
ecdsa==0.19.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jinxed==1.3.0
langdetect==1.0.9
lxml==5.3.0
numpy==2.1.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyOpenSSL==24.3.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
redis==3.5.3
requests==2.32.3
rsa==4.9
service-identity==24.2.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
zope.interface==7.2